backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.3.0
requests>=2.28.0
qrcode
